Pass functionality to generate questions into quiz
