+++ src/IntervalExercise.js
@@ -57,6 +57,30 @@ const INTERVAL_DISPLAY_INFO = {
     },
 }
 
+const POSSIBLE_INTERVALS = [
+    Teoria.interval('m2'),
+    Teoria.interval('M2'),
+    Teoria.interval('m3'),
+    Teoria.interval('M3'),
+    Teoria.interval('P4'),
+    Teoria.interval('d5'),
+    Teoria.interval('P5'),
+    Teoria.interval('m6'),
+    Teoria.interval('M6'),
+    Teoria.interval('m7'),
+    Teoria.interval('M7'),
+    // TODO: include octave
+
+]
+
+function getRandomInt(max) {
+    return Math.floor(Math.random() * Math.floor(max));
+}
+
+export function generateRandomInterval() {
+    return POSSIBLE_INTERVALS[getRandomInt(POSSIBLE_INTERVALS.length)]
+}
+
 class IntervalExercise extends Component {
     constructor() {
         super()
